// src/utils/interfaces.ts
export interface IFormFilds {
  name: string;
  age: number;
  country: string;
  email: string;
  password: string;
  confirmPassword: string;
  gender: string;
  image: FileList;
  tandc: boolean;
}

export interface IFormData {
  name: string;
  age: number;
  country: string;
  email: string;
  password: string;
  confirmPassword: string;
  gender: string;
  image: string;
  tandc: boolean;
}